Move HasQuestionnaireAnswers helper methods to separate class
