+++ decidim-forms/app/controllers/decidim/forms/admin/concerns/has_questionnaire_answers.rb
@@ -19,10 +19,13 @@ module Decidim
 
           included do
             include Decidim::Paginable
+            include Decidim::Forms::Admin::Concerns::HasQuestionnaireAnswersHelpers
 
             helper Decidim::Forms::Admin::QuestionnaireAnswersHelper
 
-            helper_method :questionnaire_url, :questionnaire_participants_url, :questionnaire_participant_answers_url, :questionnaire_export_url, :questionnaire_export_response_url
+            helper_method :questionnaire_url, :questionnaire_participants_url,
+                          :questionnaire_participant_answers_url, :questionnaire_export_url,
+                          :questionnaire_export_response_url
             helper_method :prev_url, :next_url, :first?, :last?
 
             def index
@@ -64,32 +67,6 @@ module Decidim
               raise "#{self.class.name} is expected to implement #questionnaire_for"
             end
 
-            # You can implement this method in your controller to change the URL
-            # where the questionnaire can be edited.
-            def questionnaire_url
-              url_for(questionnaire.questionnaire_for)
-            end
-
-            # You can implement this method in your controller to change the URL
-            # where the questionnaire participants' info will be shown.
-            def questionnaire_participants_url
-              url_for([:index, questionnaire.questionnaire_for, format: nil])
-            end
-
-            # You can implement this method in your controller to change the URL
-            # where the user's questionnaire answers will be shown.
-            def questionnaire_participant_answers_url(session_token)
-              url_for([:show, questionnaire.questionnaire_for, session_token: session_token])
-            end
-
-            def questionnaire_export_url
-              url_for([:export, questionnaire.questionnaire_for, format: "pdf"])
-            end
-
-            def questionnaire_export_response_url(session_token)
-              url_for([:export_response, questionnaire.questionnaire_for, session_token: session_token, format: "pdf"])
-            end
-
             private
 
             def i18n_scope
@@ -119,37 +96,6 @@ module Decidim
 
             def participants(query)
               query.map { |p| participant(p.session_token) }
-            end
-
-            # Custom pagination methods
-            def participant_ids
-              @participant_ids ||= collection.pluck(:session_token)
-            end
-
-            def current_idx
-              participant_ids.index(params[:session_token])
-            end
-
-            def prev_url
-              return if first?
-
-              token = participant_ids[current_idx - 1]
-              questionnaire_participant_answers_url(token)
-            end
-
-            def next_url
-              return if last?
-
-              token = participant_ids[current_idx + 1]
-              questionnaire_participant_answers_url(token)
-            end
-
-            def first?
-              current_idx.zero?
-            end
-
-            def last?
-              current_idx == participant_ids.count - 1
             end
           end
         end
